requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
bidict==0.22.0
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.3
multidict==6.0.2
numpy==1.23.2
PyAudio==0.2.12
pydub==0.25.1
pygame==2.1.2
pyserial==3.5
python-engineio==4.3.4
python-socketio==5.7.1
yarl==1.8.1
